Refactor `MiddlewarePlugin`
