+++ packages/next/server/web/sandbox/context.ts
@@ -10,7 +10,7 @@ import {
   AbortSignal,
 } from 'next/dist/compiled/abort-controller'
 import vm from 'vm'
-import type { WasmBinding } from '../../../build/webpack/loaders/next-middleware-wasm-loader'
+import type { WasmBinding } from '../../../build/webpack/loaders/get-module-build-info'
 
 const WEBPACK_HASH_REGEX =
   /__webpack_require__\.h = function\(\) \{ return "[0-9a-f]+"; \}/g
